Module austenite::content_neg

Content Negotiation

Functions

best_charset
Return the best allowed charset.

best_content_type
Return the best allowed content type for the request. The best type is the first type in the Accept header that is available.

best_encoding
Return the best allowed encoding.

best_language
Return the best allowed language.